requirements.txt
flask>=2.3.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
pinecone>=2.2.0
colorama>=0.4.6
bcrypt>=4.0.0
flask-talisman>=1.1.0
flask-limiter>=2.8.0
gunicorn>=21.0.0 